fix: change to `update -w` per feedback
