+++ src/cargo/ops/cargo_generate_lockfile.rs
@@ -17,6 +17,7 @@ pub struct UpdateOptions<'a> {
     pub precise: Option<&'a str>,
     pub aggressive: bool,
     pub dry_run: bool,
+    pub workspace: bool,
 }
 
 pub fn generate_lockfile(ws: &Workspace<'_>) -> CargoResult<()> {
@@ -35,6 +36,19 @@ pub fn generate_lockfile(ws: &Workspace<'_>) -> CargoResult<()> {
 }
 
 pub fn update_lockfile(ws: &Workspace<'_>, opts: &UpdateOptions<'_>) -> CargoResult<()> {
+    if opts.workspace {
+      if opts.aggressive {
+        anyhow::bail!("cannot specify aggressive for workspace updates");
+      }
+      if opts.precise.is_some() {
+        anyhow::bail!("cannot specify precise for workspace updates");
+      }
+
+      ws.emit_warnings()?;
+      let (_packages, _resolve) = ops::resolve_ws(ws)?;
+      return Ok(())
+    }
+
     if opts.aggressive && opts.precise.is_some() {
         anyhow::bail!("cannot specify both aggressive and precise simultaneously")
     }
