+++ src/cargo/ops/mod.rs
@@ -15,7 +15,6 @@ pub use self::cargo_package::{package, PackageOpts};
 pub use self::cargo_pkgid::pkgid;
 pub use self::cargo_read_manifest::{read_package, read_packages};
 pub use self::cargo_run::run;
-pub use self::cargo_sync_lockfile::{sync_lockfile, SyncLockfileOptions};
 pub use self::cargo_test::{run_benches, run_tests, TestOptions};
 pub use self::cargo_uninstall::uninstall;
 pub use self::fix::{fix, fix_maybe_exec_rustc, FixOptions};
@@ -42,7 +41,6 @@ mod cargo_package;
 mod cargo_pkgid;
 mod cargo_read_manifest;
 mod cargo_run;
-mod cargo_sync_lockfile;
 mod cargo_test;
 mod cargo_uninstall;
 mod common_for_install_and_uninstall;
