+++ src/doc/man/cargo-update.md
@@ -44,6 +44,14 @@ the package to. If the package comes from a git repository, this can be a git
 revision (such as a SHA hash or tag).
 {{/option}}
 
+{{#option "`-w`" "`--workspace`" }}
+Attempt to update only packages defined in the workspace. Other packages
+are updated only if they don't already exist in the lockfile. This
+option is useful for updating `Cargo.lock` after you've changed version
+numbers in `Cargo.toml`.
+Cannot be used with `--precise` or `--aggressive`.
+{{/option}}
+
 {{#option "`--dry-run`" }}
 Displays what would be updated, but doesn't actually write the lockfile.
 {{/option}}
