+++ src/doc/man/generated_txt/cargo-update.txt
@@ -35,6 +35,13 @@ OPTIONS
            to set the package to. If the package comes from a git repository,
            this can be a git revision (such as a SHA hash or tag).
 
+       -w, --workspace
+           Attempt to update only packages defined in the workspace. Other
+           packages are updated only if they don't already exist in the
+           lockfile. This option is useful for updating Cargo.lock after you've
+           changed version numbers in Cargo.toml. Cannot be used with --precise
+           or --aggressive.
+
        --dry-run
            Displays what would be updated, but doesn't actually write the
            lockfile.
